Introducing extensions for memcpy family of memory operations
This commit introduces two new extensions to support the new instructions
provided on the 2021 Arch Extensions for standardising the memcpy family
of memory operations.
This includes a new feature macro, `__ARM_FEATURE_MOPS`, which signals
the availability of the instructions, and a new intrinsic covering the
memset with memory tagging operation.