[github.io] Add markdown-link-check as part of CI (#43)
The two CI jobs (`build` and `markdown-link-check`) run in parallel but that doesn't matter to the end goal of the markdown link check, because the `build` step checks that the new files generated via the scripts from the intrinsic DB are same as the one in the `mve_intrinsics` folder. This means that at the end the link checks will have to run on an up-to-date version of the specs otherwise the `build` step will fail anyway.