[cleanup] Turn predefined feature macros into links
The table of predefined macros used links of the form “[[ssec-FOO]]”
after the macro name to link to the associated documentation.
The “FOO” was usually quite cryptic and wasn't likely to make
immediate sense to the reader.  This patch converts the macros
themselves into links instead.

Some macros had multiple links.  In those cases the patch uses
the first link target for the macro and continues to list the
other links separately.  Later patches clean this up, so that
each macro has only one link.