[cleanup] Link directly to specific crypto sections
The table of feature macros used links to “Crypto extension” as the
main link target for each crypto macro, followed by a link to the
feature-specific section.  The latter link seems more useful than the
former, since the documentation of the crypto extension starts with:

  NOTE: The `__ARM_FEATURE_CRYPTO` macro is deprecated in favor of the
  finer grained feature macros described below.

If the reader is already looking at one of the finer-grained
macros then we might as well take them directly to the associated
section.