[cleanup] Fix link target for __ARM_FEATURE_SIMD32
__ARM_FEATURE_SIMD32 links first to “Saturation instructions”,
but that section doesn't reference the macro.  It seems better
to link to “32-bit SIMD instructions” instead.  That in turn
has a link to #32-bit-simd-operations, which was the row's
other link target.

The default name for the link target is #bit-simd-instructions,
so the patch uses an explicit <span> to get a better name.