[cleanup] Prefer “availability of” links
Some entries in the feature macro table linked first to the description
of the macro itself, and then to a separate “availability of” section.
The latter sections link to the former, so this patch makes them the
target of the table entries' main link.

It feels a bit inconsistent to have this structure only for three
macros, so perhaps more refactoring is needed.  In the menatime,
it seemed better to link there than to lose the link entirely.