[cleanup] Rework intrinsics section
The “Intrinsics” section first describes Neon intrinsics, then has a
paragraph about MVE intrinsics, and then has a paragraph about “other”
intrinsics.  The MVE part is an adjusted cut-&-paste of the Neon part,
so IMO this doesn't read very naturally.  There's also a lingering
assumption that arm_acle.h is for all “non-Neon“ intrinsics,
which is no longer true.

Also, the Neon and MVE text include links that would be useful
in the associated header files section.

This patch therefore restricts the “Intrinsics” section to talking
about intrinsics in general, with a link to the “Header files”
section for individual groups of intrinsics.  It copies across
the link targets mentioned above.