[cleanup] Add subsections for each header file
The “header files” section first described arm_acle.h and
arm_neon.h, then went on to describe things that apply to
ACLE header files in general.  The list of header files
was then expanded to include arm_mve.h, arm_fp16 and arm_bf16.h.

I found this made the section hard to read.  The description of
each header file was essentially cut-&-pasted from arn_neon.h
and adjusted, so the text didn't flow naturally as a single
section.  It was also harder to tell that the text at the end
of section applied to all header files mentioned earlier.

This patch restructures the section so that it first describes
the common features of ACLE header files and then has a separate
subsection for each individual header file.