[main] Add support for SVE
This patch incorporates the main text of the SVE ACLE, moving
the SVE ACLE from “beta” to “release” status at the same time.
The separate beta document is no longer going to be maintained.

For now the text refers to the interactive intrinsics on
developer.arm.com for the full list of intrinsics.  However,
it does include the (hand-written) table that maps SVE instructions
to their associated intrinsics, with the intrinsics again being
links to developer.arm.com.  This is often more precise than
it was in the separate beta doc, since the interactive intrinsics
tool allows the functions to be filtered by element size or data type.
However, there are some cases in which it isn't possible to narrow
the search term down exactly.  Examples include:

* Contiguous extending loads, where an explicit suffix is always
  required.  It is not possible to include those suffixes while
  excluding _gather.

* Stores, where the element size filter has no effect.
  (The filter applies to the return type, which is always
  void for stores.)

The new links should be better on balance though.

For now the instruction-to-intrinsic table includes only the
FEAT_SVE instructions.  A later patch will add the rest.

Other notes:

Quoting a table (as for the table with a “Type category” heading)
caused pandoc with a default LaTeX template to blow up with:

---------------------------------------------------------------
! pdfTeX error (ext4): \pdfendlink ended up in different nesting level than \pd
fstartlink.
\AtBegShi@Output ...ipout \box \AtBeginShipoutBox
                                                  \fi \fi
l.8921 \end{quote}

!  ==> Fatal error occurred, no output PDF file produced!
---------------------------------------------------------------

But it works (and produces correctly-formatted output) with our
template.  The patch therefore makes hyperref-check.sh use the
template too.

Things to do separately:

* Replace existing non-SVE uses of “vector” with something more
  specific, in cases where “vector” refers to Advanced SIMD vectors
  in particular.

* List other SVE instructions, not just the FEAT_SVE ones.

* Improve the formatting of level 5 headers in the PDF output.

* Much longer term: create a separate document that lists the
  SVE intrinsics, rather than relying on developer.arm.com.

Co-authored-by: sallyarmneale <56446080+sallyarmneale@users.noreply.github.com>