[main] Add instructions for SVE extensions
The cover note for the original SVE patch said:

    For now the instruction-to-intrinsic table includes only the
    FEAT_SVE instructions.  A later patch will add the rest.

This patch does that.  It also tightens some existing intrinsic
links to exclude unrelated functions.  For example, linking to
svcvt_ excludes svcvtlb and svcvtlt.

The patch doesn't add an entry to the changelog since it is
really just completing the existing SVE work, which is already
listed as a change in the next release.