[main] Add an _m suffix to some SME intrinsics
While implementing the SME ACLE for GCC, I noticed that
svwrite was inconsistent with svadd* and with the MOP
intrinsics.  svwrite (which maps to MOVA) had an _m suffix,
to reflect the /M on the assembly predicate operands, but the
other functions didn't.

We could resolve the inconsistency in either direction:
remove _m from svwrite or add it to the other intrinsics.
Although I don't have a strong preference, adding the _m
to the other intrinsics seems more future proof.  SME2 adds
various unpredicated intrinsics and there's a risk of a future
name clash if we don't establish a distinction between /m
functions and unpredicated functions.  Making _m explicit
also seems more consistent with SVE.