docs: use GitHub absolute links in PR and issue templates
GitHub does not work correctly with relative links in PR and issue
templates. Because of this, wrong links are generated in Pull Requests
and Issues (i.e. error 404).

As far as I could find, there is no other solution besides changing them
to absolute links. This is unfortunate because we now link directly to
GitHub pages, but it's better than showing error 404 to contributors.