Add SLC as a valid Prefetch Target
Arm v8.9-a/v9.4-a has allocated the encoding 0b11 in PRFM's `Rt<2:1>`
field (which represents the cache level targetted by the prefetch
instruction) the name "SLC", to represent prefetches to the system-level
cache. In some documents this is called FEAT_PRFMSLC.

This change updates the ACLE intrinsics to match what is available in
the PRFM instructions.