Clarify licensing information
One of the files under the tools directory is under the BSD 3-Clause
license, but this fact is not clear in our licensing descriptions. This
has been fixed.

Additionally, a copy of the BSD 3-Clause is included in the project in
order to abide by its requirements.

This patch also fixes some related typos and changes all instances of
"licence" to the US English spelling.