Remove incorrect optimisation remark in CMSE's floating-point register clearing
The section "Arguments on the stack and floating point handling"
contains an incorrect remark about its example code sequence.

It states that clearing FP registers only if they contain Secure
information is an optimisation in the sense that it can be done
regardless. This is wrong since the clearing must not be done if they
don't have Secure information. Otherwise an unnecessary FP context would
be created, which could have serious side effects such as the raise of
UsageFault.

This patch removes the incorrect text.