Correct description and example in CMSE's Non-secure function pointers
The original description and example incorrectly showed that a
non-secure function pointer must be cast into a regular function pointer
if the target function wasn't non-secure. This is incorrect because the
corresponding call instruction, BLXNS, is able to handle calls that
don't involve security state transitions. In other words, it
supports target functions that are either secure or non-secure.

This patch corrects the description and the example, and emphasizes that
telling apart secure and non-secure target functions in code is simply a
matter of optimisation and is not required.