Use keywords for SME attributes
A patch series recently went into Clang to support a kind of keyword
attribute that can appear wherever a standard attribute can appear,
and that appertains to whatever a standard attribute appertains to:
https://reviews.llvm.org/D148700

This patch changes the syntax used for SME to match this style, rather
than using GNU attributes.  That avoid the previous situation in which
older compilers would ignore (or just warn about) SME attributes,
and so would generate wrong code.