Add ACLE macros for FEAT_PAuth_LR
The Armv9.5a introduces the FEAT_PAuth_LR extension, adding an enhancement
to the Pointer Authentication extension where the value of PC is used as
a diversifier when signing pointers.

This updates the ACLE to support the extension by the introduction of a
new feature marcro and the addition of a new possible value to the
existing `__ARM_FEATURE_PAC_DEFAULT` macro.