[PDF] Fix searching and copy-pasting underscore characters
In the existing setup, the *.md source files are converted to PDF by
pandoc that invokes pdflatex internally. With the default font encoding,
underscore characters inside paragraphs of text look like whitespace
(or absent) in the produced PDF documents w.r.t. copy-pasting text from
PDF viewer or searching. This may confuse users as it makes
`__ARM_FEATURE_name` and `long_function_name` strings invisible to the
"Search ..." function of a viewer, but only if they are not inside a
standalone block of code.

One of the solutions is to use T1 font encoding and ensure that Type 1
fonts are available (i.e. pdflatex does not have to use rasterized
Type 3 fonts).