Remove comments before <span>s (#334)
ACLE had developed a style of prefixing empty <span>s with:

<!-- Do not remove the following `span`, it is needed to create an
anchor that can be referred via an internal hyperlink to the paragraph
following it. -->

However, this appears to confuse the GitHub renderer when used
in bullet points.  It was skipping large portions of text after
the streaming-intrinsic and vl-dependent anchors, and formatting
later text as though it was part of the same bullet point.

Also, I think these large comments make it harder to see the <span>
itself.  Probably because of this,

   16-bit-floating-point-data-processing-operations

had become unmoored from the associated text.

This patch removes the comments (which seems to fix the rendering)
and moves 16-bit-floating-point-data-processing-operations to the
appropriate place.  I checked each anchor manually to make sure
that it still worked.