Include SME attributes in the name mangling of types
This change extends the name mangling of types to include the SME streaming
and ZA interface. This will avoid naming conflicts which can currently
arise such as in the following example:

  void foo(void (*f)()) { f(); }
  void foo(void (*f)() __arm_streaming) { f(); }

Without this change, both functions 'foo' above will mangle to the same
name, despite the function pointers being different.