Add clarification and minor changes to FMV scope and signatures
Here I add some clarifications about FMV for target versions.

In particular, this changes the rules to make the default function
the definitive version for signatures and scope of the function.

This means that the default version declaration is used for defining
default argument values, and is required to be in scope for any call
to the function set.