[SME] Add __arm_agnostic("sme_za_state") keyword attribute (#336)
The `__arm_agnostic` keyword attribute enables the user to specify that
a function is agnostic to a specified piece of architectural state. That
means that the function must preserve this state when it exists, or
otherwise ignores its contents.

The reason for not naming this something like `__arm_za_compatible` was
so that we might want use the attribute keyword for other architectural
state in the future.